requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
groq
cassio
beautifulsoup4
langchain-groq
wikipedia
arxiv
langchainhub
sentence_transformers
PyPDF2
langchain-objectbox
pyngrok
pytest
flask
ollama
langchain
pdfplumber
pymupdf 
google-search-results
ollama
huggingface_hub
llama-cpp-python==0.1.78
numpy==1.23.4